feat: pegar eventos pela data
