+++ src/consts/queryKeys.js
@@ -2,6 +2,7 @@ const QUERY_KEYS = {
 	DOCENTS: "docents",
 	COUNTERS: "counters",
 	EVENTS: "events",
+	EVENTS_BY_DATE: "eventsByDate",
 };
 
 export default QUERY_KEYS;
